🔼 dependency bumps (incl braeking changes)
