--- cypress.json
@@ -1,11 +1,0 @@
-{
-	"baseUrl": "http://localhost:3000/#",
-	"defaultCommandTimeout": 60000,
-	"requestTimeout": 60000,
-	"video": true,
-	"videoCompression": false,
-	"screenscreenshotOnRunFailureshots": true,
-	"chromeWebSecurity": false,
-	"watchForFileChanges": false
-}
-
