+++ cypress/e2e/user-admin/1-welcome-page.js
