+++ cypress/e2e/user-claimer/1-view-and-claim.js
