+++ cypress/e2e/user-claimer/2-view-and-claim-w-game.js
@@ -3,6 +3,7 @@
 // /////////////////////////////*/
 
 const admin = require( '../../fixtures/admin-user' )
+const oneCode = require( `../../fixtures/one-correct-code${ Cypress.env('LOCAL') ? '' : '-ci' }` )
 const request_options = {
 
 	headers: {
@@ -24,7 +25,7 @@ async function extract_challenge_from_url ( response ) {
 
 }
 
-context( 'User can claim POAP after succeeding at challenge game', () => {
+context( 'Claimer can view valid events with game', () => {
 
 	/* ///////////////////////////////
 	// First event
@@ -46,7 +47,7 @@ context( 'User can claim POAP after succeeding at challenge game', () => {
 
 		// Select anti-farming timing (10s)
 		cy.get( '#event-create-game-duration' ).select( 1 )
-		cy.log( 'Game time selected: 10s AKA 2 game turns' )
+		cy.log( 'Game time selected: 20s AKA 4 game turns' )
 
 		// Create event
 		cy.get( '#event-create-submit' ).click()
@@ -73,7 +74,7 @@ context( 'User can claim POAP after succeeding at challenge game', () => {
 
 	} )
 
-	it( 'Event 1: Succesfully redirects to challenge link and fail at game', function( ) {
+	it( 'Event 1: Successfully redirects to challenge link and play game', function( ) {
 
 		// Visit the public link with games
 		cy.request( { ...request_options, url: `${ Cypress.env( 'REACT_APP_publicUrl' ) }/claim/${ this.event_1_public_auth_link }` } ).as( `request` )
@@ -99,19 +100,41 @@ context( 'User can claim POAP after succeeding at challenge game', () => {
 				cy.contains( 'Play a game' )
 
 				// Click start game button
-				cy.contains( 'button', 'Start game' ).click()
+				cy.contains( 'a', 'Start game' ).click()
 
 				// Expect score text
 				cy.contains( 'Score: 0 of' )
 
+				// Find buttons in page and click the one with h1 span value
+				cy.get( 'h1 span' ).invoke( 'text' ).then( (text) => { cy.contains( 'a', text ).click( { force: true } ) })
+
+				// Expect score text
+				cy.contains( 'Score: 1 of' )
+
+				// Find buttons in page and click the one with h1 span value
+				cy.get( 'h1 span' ).invoke( 'text' ).then( (text) => { cy.contains( 'a', text ).click( { force: true } ) })
+
+				// Expect score text
+				cy.contains( 'Score: 2 of' )
+
 				// Expect winning screen
-				cy.contains( 'You lost!' )
+				cy.contains( 'You won!' )
+
+				// Click retrieval link
+				cy.contains( 'a', 'Claim your' ).click()
+
+				// Wait for code retrieval
+				cy.url().should( 'include', 'app.poap.xyz' )
+
+				// Check if POAP link supplies one of the test codes
+				cy.url().should( 'include', oneCode )
 
 			} )
-	
+		
+
 	} )
 
-	it( 'Event 1: Shows code not marked as used after failing game', function( ) {
+	it( 'Event 1: Shows code marked as used (previous redirect marked as used)', function( ) {
 
 		// Visit the public link
 		cy.visit( this.event_1_publiclink )
@@ -120,10 +143,73 @@ context( 'User can claim POAP after succeeding at challenge game', () => {
 		cy.get( '#event-view-accept-disclaimer' ).click()
 
 		// Shows one code as claimed
-		cy.contains( '0 of 1 codes' )
+		cy.contains( '1 of 1 codes' )
 
 	} )
 
+	it( 'Event 1: Previous challenge link no longer works', function( ) {
+
+		// Visit the public link
+		cy.visit( this.event_1_first_challenge_url )
+
+		// Interface should indicate that the link expired
+		cy.contains( 'This link was already used' )
+
+	} )
+
+	it( 'Event 1: Shows no codes after code is scanned', function( ) {
+
+		// Visit public event link
+		cy.visit( this.event_1_publiclink )
+
+		// Accept disclaimer
+		cy.get( '#event-view-accept-disclaimer' ).click()
+
+		cy.contains( '1 of 1 codes' )
+
+	} )
+
+	it( 'Event 1: Shows error if link was used after code ran out', function( ) {
+
+		// Visit the public link to the second code as read by simulating a scan
+		cy.request( { ...request_options, url: `${ Cypress.env( 'REACT_APP_publicUrl' ) }/claim/${ this.event_1_public_auth_link }` } ).as( `request` )
+			.then( extract_challenge_from_url )
+			.then( event_1_second_challenge => {
+
+				// Visit the challenge link
+				cy.visit( `/claim/${ event_1_second_challenge }` )
+
+				// Human game welcome screen 
+				cy.contains( 'Play a game' )
+
+				// Click start game button
+				cy.contains( 'a', 'Start game' ).click()
+
+				// Expect score text
+				cy.contains( 'Score: 0 of' )
+
+				// Find buttons in page and click the one with h1 span value
+				cy.get( 'h1 span' ).invoke( 'text' ).then( (text) => { cy.contains( 'a', text ).click( { force: true } ) })
+
+				// Expect score text
+				cy.contains( 'Score: 1 of' )
+
+				// Find buttons in page and click the one with h1 span value
+				cy.get( 'h1 span' ).invoke( 'text' ).then( (text) => { cy.contains( 'a', text ).click( { force: true } ) })
+
+				// Expect score text
+				cy.contains( 'Score: 2 of' )
+
+				// Expect winning screen
+				cy.contains( 'You won!' )
+
+				cy.on('window:alert', response => {
+					expect( response ).to.contain( 'No more POAPs available for this event!' )
+				} )
+
+			} )
+
+	} )
 
 	// Delete event 1
 	it( 'Event 1: Deletes the event when clicked', function() {
