+++ cypress/e2e/user-language/1-view-and-switch-language.js
