+++ functions/package.json
@@ -20,10 +20,10 @@
     "@sendgrid/mail": "^7.7.0",
     "body-parser": "^1.20.0",
     "cors": "^2.8.5",
-    "csso": "^5.0.3",
+    "csso": "^5.0.4",
     "express": "^4.18.1",
-    "firebase-admin": "^10.2.0",
-    "firebase-functions": "^3.21.2",
+    "firebase-admin": "^11.0.0",
+    "firebase-functions": "^3.22.0",
     "isomorphic-fetch": "^3.0.0",
     "juice": "^8.0.0",
     "papaparse": "^5.3.2",
