+++ package.json
@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "cypress": "^9.7.0",
-    "firebase": "^9.8.2",
-    "i18next": "^21.8.5",
+    "cypress": "^10.3.0",
+    "firebase": "^9.9.0",
+    "i18next": "^21.8.14",
     "i18next-browser-languagedetector": "^6.1.4",
     "i18next-http-backend": "^1.4.1",
     "papaparse": "^5.3.2",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-google-recaptcha": "^2.1.0",
-    "react-i18next": "^11.16.9",
+    "react-i18next": "^11.18.1",
     "react-qr-code": "^2.0.7",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
@@ -22,8 +22,8 @@
   "scripts": {
     "start": "DISABLE_ESLINT_PLUGIN=true react-scripts start",
     "build": "react-scripts build",
-    "cypress": "NODE_ENV=development $(npm bin)/cypress open --browser chrome",
-    "test": "CYPRESS_LOCAL=true NODE_ENV=development start-server-and-test 'BROWSER=none npm start' 3000 \"$(npm bin)/cypress open --browser chrome\"",
+    "cypress": "NODE_ENV=development $(npm bin)/cypress open --e2e --browser chrome",
+    "test": "CYPRESS_LOCAL=true NODE_ENV=development start-server-and-test 'BROWSER=none npm start' 3000 \"$(npm bin)/cypress run --browser chrome\"",
     "test:production": "npm run build && NODE_ENV=production start-server-and-test 'serve build' 3000 \"$(npm bin)/cypress open --browser chrome\"",
     "test:ci": "NODE_ENV=production CI=true start-server-and-test 'BROWSER=none npm start' 3000 \"$(npm bin)/cypress run --browser chrome --config 60000\"",
     "eject": "react-scripts eject"
@@ -47,12 +47,12 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.18.2",
+    "@babel/core": "^7.18.6",
     "@babel/eslint-parser": "^7.18.2",
-    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-react": "^7.18.6",
     "cypress-file-upload": "^5.0.8",
     "dotenv": "^16.0.1",
-    "serve": "^13.0.2",
+    "serve": "^14.0.1",
     "start-server-and-test": "^1.14.0"
   }
 }
